# doc/api.yaml
...
openapi: 3.0.3
info:
    title: 'WASAText API'
    description: |-
        A modern chat application API that allows users to create groups,
        start conversations, send messages, react to messages, and manage
        contacts.

        Features include:
        - Real-time messaging with WebSocket support
        - Image messaging with automatic compression
        - Message reactions with toggle behavior (prevents duplicates)
        - Group chat management
        - Contact management
        - Message forwarding (including images)
        - User profiles with photos
    version: '1.2.0'
security:
    - bearerAuth: []

# Note: All authenticated endpoints (those requiring bearerAuth) will return 
# a 401 Unauthorized response if authentication fails. See the UnauthorizedError 
# response definition in components for details.
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: UUID
            description: |-
                Bearer token authentication using User UUID as the token.
                The token should be the UUID of an existing user in the system.
    schemas:
...
        Error:
            type: object
            properties:
                message:
                    type: string
                    example: 'Invalid input provided'
                    description: Error message describing what went wrong
            required:
                - message
    responses:
        UnauthorizedError:
            description: Authentication information is missing or invalid
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
                    examples:
                        missing_header:
                            summary: Missing Authorization header
                            value:
                                message: "Authorization header required"
                        invalid_format:
                            summary: Invalid authorization format
                            value:
                                message: "Invalid authorization format. Use 'Bearer <token>'"
                        invalid_token:
                            summary: Invalid token format
                            value:
                                message: "Invalid token format"
                        user_not_found:
                            summary: User not found
                            value:
                                message: "Invalid token - user not found"
paths:
...
    /users/{id}:
        parameters:
            - name: id
              in: path
              description: UUID of the user
              required: true
              schema:
                  type: string
                  format: uuid
        put:
            tags: ['Users']
            summary: Update username
            description: Set or update the user's username
            operationId: setMyUserName
            requestBody:
                description: New username as a JSON string
                content:
                    application/json:
                        schema:
                            type: string
                            minLength: 3
                            maxLength: 16
                            example: 'Alice'
                required: true
            responses:
                '200':
                    description: Username updated successfully
                '400':
                    description: Invalid username format
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '401':
                    $ref: '#/components/responses/UnauthorizedError'
                '404':
                    description: User not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /users/{id}/photo:
        parameters:
            - name: id
              in: path
              description: UUID of the user
              required: true
              schema:
                  type: string
                  format: uuid
        put:
            tags: ['Users']
            summary: Update user photo
            description: Set or update the user's profile picture
            operationId: setMyPhoto
            requestBody:
                description: New photo URL as a JSON string
                content:
                    application/json:
                        schema:
                            type: string
                            example: 'https://example.com/new-avatar.jpg'
                required: true
            responses:
                '200':
                    description: Photo updated successfully
                '400':
                    description: Invalid photo URL
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: User not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
...